tool_repository/requirements.txt
click==8.0.1
Flask==2.0.1
greenlet==1.1.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysqlclient==2.0.3
PyMySQL==1.0.2
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.23
SQLAlchemy-Utils==0.37.8
Werkzeug==2.0.1
